+++ apps/frontend-user/.gitignore
@@ -1,3 +1,18 @@
-dist
-dist-ssr
-.vite
+# Local
+.DS_Store
+*.local
+*.log*
+
+# Dist
+node_modules
+dist/
+.vinxi
+.output
+.vercel
+.netlify
+.wrangler
+
+# IDE
+.vscode/*
+!.vscode/extensions.json
+.idea
