--- apps/frontend-user/frontend-user/.gitignore
@@ -1,18 +1,0 @@
-# Local
-.DS_Store
-*.local
-*.log*
-
-# Dist
-node_modules
-dist/
-.vinxi
-.output
-.vercel
-.netlify
-.wrangler
-
-# IDE
-.vscode/*
-!.vscode/extensions.json
-.idea
