+++ apps/frontend-user/package.json
@@ -1,35 +1,26 @@
 {
-  "name": "@mindseed/frontend-user",
-  "private": true,
+  "name": "frontend-user",
   "version": "0.0.0",
+  "private": true,
   "type": "module",
   "scripts": {
-    "dev": "vite",
-    "build": "tsc -b && vite build",
-    "lint": "eslint .",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "@mindseed/api-types": "workspace:*",
-    "@tanstack/react-router": "^1.167.4",
-    "react": "^19.2.0",
-    "react-dom": "^19.2.0"
+    "typecheck": "tsc --noEmit",
+    "dev": "vite --port=3001",
+    "build": "vite build",
+    "serve": "vite preview",
+    "start": "vite"
   },
   "devDependencies": {
-    "@eslint/compat": "^2.0.2",
-    "@eslint/js": "^9.39.1",
-    "@tanstack/router-devtools": "^1.166.9",
-    "@tanstack/router-plugin": "^1.166.13",
-    "@types/node": "^24.10.1",
-    "@types/react": "^19.2.7",
-    "@types/react-dom": "^19.2.3",
-    "@vitejs/plugin-react": "^5.1.1",
-    "eslint": "^9.39.1",
-    "eslint-plugin-react-hooks": "^7.0.1",
-    "eslint-plugin-react-refresh": "^0.4.24",
-    "globals": "^16.5.0",
-    "typescript": "~5.9.3",
-    "typescript-eslint": "^8.48.0",
-    "vite": "^7.3.1"
+    "@types/react": "^19.0.8",
+    "@types/react-dom": "^19.0.3",
+    "@tanstack/router-plugin": "^1.117.2",
+    "@vitejs/plugin-react": "^4.3.2",
+    "vite": "^6.0.3"
+  },
+  "dependencies": {
+    "@tanstack/react-router": "^1.117.1",
+    "@tanstack/router-devtools": "^1.117.1",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0"
   }
 }
