--- apps/frontend-user/src/App.tsx
@@ -1,7 +1,0 @@
-import "./App.css";
-
-function App() {
-  return <></>;
-}
-
-export default App;
